requirements.txt
pytest
discord.py
requests
psycopg2